Desktop indexing/search utility
Scans text files in a given folder and maintains a flatfile index
sensitive to updates. The results are displayed in decreasing order of
relevance. JS based intuitive and helpful UI. Maintains history.
Intelligent suggestions based on past events. The results shown contains
specific lines extracted from the file so that user can check the
context in which the keyword has been used.